Added "NOCLOBBER" version of FP-setting macros, needed by migration library
For callee-saved FP registers: if we mark them as clobbered in the inline
assembly when setting the FP registers post-migration, the compiler will
arrange to save/restore them in the stack.  Thus, we have to set them without
informing the compiler so the values "stick" when returning back to the
application.