More robust sorting of live values in stackmaps
Previously, if a value didn't have a name we sorted by pointer value.
However, during a multi-ISA compilation we clone the IR, meaning the
order in which values are allocated (and their corresponding locations
in the stackmap) may change.  To force an ordering, utilize LLVM's
SlotTracker machinery which assigns numbers to unnamed values.