Bug fix: restore callee-saved registers post-migration
Turns out it's difficult for the kernel to restore callee-saved
registers on the syscall return path for some architectures.  Rather
than hack the return path (which is assembly and highly optimized), add
a tiny bit of glue code in the migration library to fix this up in
userland.  This also obviates the need for the SET_FP_REGS call in
__migrate_shim_internal.

Applies to:
        Migration library, libmigrate.a