Cleanup & correctly handle callee-saved registers for homogeneous migration
If we don't switch architectures during migration, we don't need to
transform the stack but we *do* need to fix up any callee-saved
registers post-migration.  However we can't simply jump back to the
beginning of __migrate_shim_internal(), as this will add another stack
frame on the stack.  This commit handles this case.