Added skewed work distribution for statically-scheduled loops
Users can assign a core speed rating to each node specified by
POPCORN_PLACES using the same syntax in POPCORN_HET_WORKSHARE.  The
ratings are integers and specify how many more iterations are assigned
per-thread.  For example, if the user specified a rating of 1 for node
A and a 2 for node B, threads on node B will get twice as many
iterations as threads on node A.  Note that this does *not* affect how
many threads are run on each node, but simply how much work each thread
running on a particular node gets.