Bug fix: correctly handle stackmap temporaries
Depending on the backend, the live values listed in the stackmap may or
may not actually be saved for use.  For example, oftentimes for stack
references the x86-64 backend can materialize addresses as needed whereas
AArch64 backend tends to generate and save references.  The problem
arises when the backends do different things.  This patch aims to
reconcile the differences by adding an extra temporary flag to the live
value metadata so the runtime can handle differences appropriately.