Change alignments of activation metadata sections -- see below
When changing the activation metadata layout, we don't want to
accidentally run afoul of alignment restrictions, which forces the
linker to insert padding when concatenating sections from multiple
binaries and screws up all of our tools (which expect all metadata
records to be contiguous).  Instead, set the alignment to be 1 so that
the linker never inserts padding between sections.