Avoid continuing up def-chain for non-copies in duplicate location finding
By the time the machine code reaches StackTransformMetadata, it is no
longer in SSA form.  Because of this, some registers which look like
copies of other registers are actually *not* copies because there are
intervening defining instructions between the initial copy and use
which makes the register no longer a copy.  This patch finds one such
situation where the backend generated references to fields in a struct.