Use code size for range checking when searching for functions
There might be duplicate function records caused by static/template
functions in headers that get resolved to the same function by the
linker.  Previously, searching for function records used the address of
the subsequent function record as the upper bound for the current
function, but duplicates cause this logic to fail.